auditron/requirements.txt
# requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
fastapi_mcp 

# Cloud SDKs
boto3
google-cloud-storage


azure-identity
azure-mgmt-storage
azure-mgmt-resource
azure-mgmt-sql
azure-mgmt-network
azure-mgmt-monitor
azure-mgmt-security